+++ project-2324-rust-std/src/prim_naive.rs
@@ -5,14 +5,14 @@ use std::collections::HashSet;
 
 /// Naive implementation of prim's algorithm as given in the subject
 /// Use of linear schearch
-pub fn prim_naive_function(map: &Map) -> (Vec<i32>, i32) {
+pub fn prim_naive_function(map: &Map) -> (Vec<i64>, i64) {
     // Initialize all the variables
     let maze_size: usize = map.width * map.height;
-    let mut cheapest: Vec<i32> = vec![i32::MAX; maze_size];
-    let mut predecessor: Vec<i32> = vec![-1; maze_size];
-    let mut capitalize_n: HashSet<i32> = (0..maze_size as i32).collect();
+    let mut cheapest: Vec<i64> = vec![i64::MAX; maze_size];
+    let mut predecessor: Vec<i64> = vec![-1; maze_size];
+    let mut capitalize_n: HashSet<i64> = (0..maze_size as i64).collect();
     // We chose the node 0;
-    let mut v: i32 = 0;
+    let mut v: i64 = 0;
     cheapest[v as usize] = 0;
 
     while !capitalize_n.is_empty() {
@@ -21,8 +21,8 @@ pub fn prim_naive_function(map: &Map) -> (Vec<i32>, i32) {
             .iter()
             .min_by_key(|&node| cheapest[*node as usize])
             .unwrap();
-        /* let mut arg_min: i32 = *capitalize_n.iter().next().unwrap();
-        let mut min: i32 = cheapest[arg_min as usize];
+        /* let mut arg_min: i64 = *capitalize_n.iter().next().unwrap();
+        let mut min: i64 = cheapest[arg_min as usize];
 
         for vertex in capitalize_n.iter() {
             if cheapest[*vertex as usize] < min {
